refactor(tasks): move Go2 rough owner [skip ci]
