+++ src/unilab/envs/locomotion/go2/__init__.py
@@ -1,2 +1,1 @@
 from .joystick import Go2JoystickCfg, Go2WalkTask
-from .rough import Go2JoystickRoughCfg, Go2JoystickRoughEnv
