+++ src/unilab/tasks/__init__.py
@@ -11,6 +11,7 @@ __unilab_registry_modules__ = (
     "unilab.envs.locomotion.go1",
     "unilab.tasks.locomotion.go2.footstand",
     "unilab.envs.locomotion.go2",
+    "unilab.tasks.locomotion.go2.rough",
     "unilab.envs.locomotion.go2w",
     "unilab.envs.locomotion.g1",
     "unilab.envs.locomotion.go2_arm",
