+++ src/unilab/tasks/locomotion/go2/__init__.py
@@ -1,3 +1,9 @@
 from .footstand import Go2FootStandCfg, Go2FootStandTask
+from .rough import Go2JoystickRoughCfg, Go2JoystickRoughEnv
 
-__all__ = ["Go2FootStandCfg", "Go2FootStandTask"]
+__all__ = [
+    "Go2FootStandCfg",
+    "Go2FootStandTask",
+    "Go2JoystickRoughCfg",
+    "Go2JoystickRoughEnv",
+]
